release 3.0.21-beta source code for java

+++ services/mpc/src/main/java/com/huaweicloud/sdk/mpc/v1/MpcMeta.java
@@ -944,7 +944,7 @@ public class MpcMeta {
         builder.withRequestField("template_id",
             LocationType.Query,
             FieldExistence.NON_NULL_NON_EMPTY,
-            Long.class,
+            Integer.class,
             f -> f.withMarshaller(DeleteTemplateRequest::getTemplateId, (req, v) -> {
                 req.setTemplateId(v);
             })
